requirements/test.txt
# Libraries for running app here.
pytest
pytest-django
pytest-cov

factory_boy
testfixtures
coverage
